+++ NEWS.md
@@ -5,6 +5,7 @@ If you are viewing this file on CRAN, please check [the latest news on GitHub](h
 - Fix: #771 read VPC files with absolute path
 - Enhance: `crs()` `is.empty()` and `area()` are now inherits from `terra` and no longer clash with `terra`.
 - Enhance: #776 `readLAScatalog` can skip corrupted files
+- Enhance: many operation are now faster on TLS data
 
 ## lidR v4.1.2 (Release date: 2024-07-09)